fix : 현재 사용자와 가장 가까운 건물 조회하는 api 호출 위치 수정 및 내 위치 공유 로직 수정
