+++ src/pages/Home/Home.tsx
@@ -10,14 +10,14 @@ import { useEffect, useState } from "react";
 import FriendLocation from "../../components/HomeContent/FriendLocation/FriendLocation";
 import MyLocationRanking from "../../components/HomeContent/MyLocationRanking/MyLocationRanking";
 import HomeNotice from "../../components/HomeContent/HomeNotice/HomeNotice";
-import { useNavigate } from "react-router-dom";
+// import { useNavigate } from "react-router-dom";
 import ShareLocationModal from "../../components/Map/ShareLocationModal/ShareLocationModal";
 import { isMyLocationInSchool } from "../../utils/mapRangeUtils";
 import { Coordinate } from "../../../types/mapTypes";
 import { checkIsSharedApi } from "../../apis/map";
 
 const Home = () => {
-  const navigate = useNavigate();
+  // const navigate = useNavigate();
   // const [showSplash, setShowSplash] = useState(true);
   const [isSharedLocation, setIsSharedLocation] = useState(false); // 내 위치 공유상태인지 아닌지
   const [hasNewAlarm, setHasNewAlarm] = useState(false); // 새로운 알람이 있는지
@@ -29,9 +29,12 @@ const Home = () => {
   // vercel 배포 시 에러 방지용
   console.log("학교 내부인지?:", isInSchool);
 
+  // 유저의 위치와 가장 가까운 위치 저장할 상태
+  const [nearLocation, setNearLocation] = useState("");
+
   // 내 위치 공유 버튼 모달 상태
   const [shareModalState, setShareModalState] = useState(false);
-  const [currenLocation, setCurrentLocation] = useState<Coordinate | null>(
+  const [currentLocation, setCurrentLocation] = useState<Coordinate | null>(
     null
   ); // 현재 위치
 
@@ -63,7 +66,7 @@ const Home = () => {
     const token = localStorage.getItem("accessToken");
     if (!token) {
       localStorage.clear();
-      navigate("/login");
+      // navigate("/login");
       return;
     }
     // 서버에 새로운 알람이 있는지 검증. 있다면 true로
@@ -91,10 +94,12 @@ const Home = () => {
             setModalState={setShareModalState}
           />
         )} */}
-        {currenLocation !== null && (
+        {currentLocation !== null && (
           <HomeMiniMap
             isSharedLocation={isSharedLocation}
             setModalState={setShareModalState}
+            currentLocation={currentLocation}
+            setNearLocation={setNearLocation}
           />
         )}
         <FriendLocation isSharedLocation={isSharedLocation} />
@@ -105,7 +110,7 @@ const Home = () => {
       <ShareLocationModal
         modalState={shareModalState}
         isSharedLocation={isSharedLocation}
-        currentLocation={currenLocation}
+        nearLocation={nearLocation}
         setModalState={setShareModalState}
         refreshSharedStatus={() =>
           setLocationSharedRefreshKey((prev) => prev + 1)
